\name{internals}
\alias{compute.var}
\alias{computeBeta}
\alias{computeMoments}
\alias{computeTiltDirect}
\alias{ComputeZ.fromS.R}
\alias{ComputeZ.fromS.R.partial}
\alias{computeZ.onechange}
\alias{computeZ.onechange.sample}
\alias{computeZ.squarewave.sample}
\alias{dchi}
\alias{fcompute.max.Z}
\alias{fmscbs}
\alias{fscan.max}
\alias{getCutoffMultisampleWeightedChisq}
\alias{matrix.max}
\alias{mscbs.classify}
\alias{pmarg.sumweightedchisq}
\alias{pvalueMultisampleWeightedChisq}
\alias{vu}
\alias{BAF2mBAF}
\alias{Mode}
\alias{check.overlap}
\alias{cnv.data.chr}
\alias{compute.baseline}
\alias{compute.baseline.manual}
\alias{delta.sd}
\alias{diagnosis.QQ.plot}
\alias{impute.missing.data}
\alias{merging.segments.chr}
\alias{seg.summary}
\alias{seq.cnv}
\alias{seq.data}
\alias{seq.segs}
\alias{seq.segs.merge}
\alias{snp.cnv}
\alias{snp.cnv.anno}
\alias{snp.cnv.refine}
\alias{snp.cnv.refine.anno}
\alias{snp.data}
\alias{snp.segs}
\alias{snp.segs.merge}
\alias{snp_table}
\alias{vcf_table}

\title{
Internal Functions and Data
}

\description{These are the functions and data to which the users do not need to directly get access.
}
