\name{saasCNV-package}
\alias{saasCNV-package}
\alias{saasCNV}
\docType{package}
\title{
Somatic Copy Number Alteration Analysis Using Sequencing and SNP Array Data
}
\description{
Perform joint segmentation on two signal dimensions derived from 
total read depth (intensity) and allele specific read depth (intensity) for 
whole genome sequencing (WGS), whole exome sequencing (WES) and SNP array data.
}
\details{
\tabular{ll}{
Package: \tab saasCNV\cr
Type: \tab Package\cr
Version: \tab 0.3.4\cr
Date: \tab 2016-05-10\cr
License: \tab GPL (>= 2)\cr
}
See the vignettes of the package for more details.
}
\author{
Zhongyang Zhang [aut, cre], Ke Hao [aut], Nancy R. Zhang [ctb]

Maintainer: Zhongyang Zhang <zhongyang.zhang@mssm.edu>
}
\references{
Zhang, Z. and Hao, K. (2015) SAAS-CNV: A joint segmentation approach on aggregated and allele specific signals 
for the identification of somatic copy number alterations with next-generation sequencing Data. 
\emph{PLoS Computational Biology}, \bold{11(11)}:e1004618.

Zhang, N. R., Siegmund, D. O., Ji, H., Li, J. Z. (2010) 
Detecting simultaneous changepoints in multiple sequences. \emph{Biometrika}, \bold{97(3)}:631--645.
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%%\keyword{ package }
\seealso{
%%\code{\link[DNAcopy:DNAcopy-package]{DNAcopy}}
\code{DNAcopy}
}
\examples{
## See the vignettes of the package for examples.
}
